+++ src/database/views.sql
@@ -86,8 +86,8 @@ FROM
 			listing_id,
 			ARRAY_AGG(JSONB_BUILD_OBJECT ('listing_update_id',
 					listing_update_id,
-					'description',
-					description,
+					'listing_update_description',
+					listing_update_description,
 					'pics',
 					pics,
 					'created_on',
@@ -104,7 +104,7 @@ FROM
 			ARRAY_AGG(JSONB_BUILD_OBJECT ('milestone_id',
 					milestone_id,
 					'milestone_description',
-					description,
+					milestone_description,
 					'date',
 					date)) milestones
 		FROM
@@ -132,7 +132,7 @@ SELECT
 	ali.participants,
 	p.nickname,
 	p.profile_picture,
-	to_tsvector(l.title || ' ' || l.category || ' ' || array_to_string(ali.locations::text [], ' ')) AS keyword_vector
+	to_tsvector(l.listing_title || ' ' || l.category || ' ' || array_to_string(ali.locations::text [], ' ')) AS keyword_vector
 FROM
 	listing l
 	LEFT JOIN agg_listing_info ali ON l.listing_id = ali.listing_id